Exerpimental 'anderson' lock, changed some API names, house cleaning

+++ src/rng.c
@@ -2,16 +2,16 @@
 // Copyright 2018 Sepehr Taghdisian (septag@github). All rights reserved.
 // License: https://github.com/septag/sx#license-bsd-2-clause
 //
+#include <time.h>
+
 #include "sx/rng.h"
 #include "sx/hash.h"
 
-#include <time.h>
 
 // This implementation is taken from: https://github.com/mattiasgustavsson/libs/blob/master/rnd.h
 // With a minor optimization in sx_rng_gen_irange
 
-// Convert a randomized uint32_t value to a float value x in the range 0.0f <= x < 1.0f. Contributed
-// by Jonatan Hedborg
+// Convert a randomized uint32_t value to a float value x in the range 0.0f <= x < 1.0f. Contributed by Jonatan Hedborg
 SX_PRAGMA_DIAGNOSTIC_PUSH()
 SX_PRAGMA_DIAGNOSTIC_IGNORED_CLANG_GCC("-Wstrict-aliasing")
 static inline float sx__rng_float_normalized(uint32_t value)
@@ -47,7 +47,7 @@ void sx_rng_seed(sx_rng* rng, uint32_t seed)
 
 void sx_rng_seed_time(sx_rng* rng)
 {
-    sx_rng_seed(rng, sx_hash_u64_to_u32((uint64_t)time(NULL)));
+    sx_rng_seed(rng, sx_hash_u64_to_u32(time(NULL)));
 }
 
 uint32_t sx_rng_gen(sx_rng* rng)
